+++ dynarag.go
@@ -12,8 +12,8 @@ import (
 	_ "github.com/golang-migrate/migrate/v4/source/file"
 	_ "github.com/lib/pq"
 
+	"github.com/Predixus/DynaRAG/internal/rag"
 	"github.com/Predixus/DynaRAG/internal/store"
-	"github.com/Predixus/DynaRAG/rag"
 	"github.com/Predixus/DynaRAG/types"
 )
 
@@ -30,7 +30,7 @@ func (c *Client) Chunk(
 ) error {
 	_, err := store.AddEmbedding(ctx, c.config.PostgresConnStr, filePath, chunk, metadata)
 	if err != nil {
-		slog.Error("Could not process embedding: %v", err)
+		slog.Error("Could not process embedding", "error", err)
 		return err
 	}
 	return nil
@@ -46,7 +46,7 @@ func (c *Client) Similar(
 
 	res, err := store.GetTopKEmbeddings(ctx, c.config.PostgresConnStr, text, k, metadata)
 	if err != nil {
-		slog.Error("Could not get top K embeddings: %v", err)
+		slog.Error("Could not get top K embeddings", "error", err)
 		return nil, err
 	}
 	return res, nil
@@ -68,7 +68,7 @@ func (c *Client) Query(
 
 	res, err := store.GetTopKEmbeddings(ctx, c.config.PostgresConnStr, query, topN, metadata)
 	if err != nil {
-		slog.Error("Could not get top K embeddings: %v", err)
+		slog.Error("Could not get top K embeddings", "error", err)
 		return err
 	}
 
@@ -84,7 +84,7 @@ func (c *Client) Query(
 
 	err = rag.GenerateRAGResponse(documents, query, writer)
 	if err != nil {
-		slog.Error("Failed to generate RAG response: ", err)
+		slog.Error("Failed to generate RAG response", "error", err)
 		return err
 	}
 	return nil
@@ -99,7 +99,7 @@ func (c *Client) PurgeChunks(ctx context.Context, dryRun *bool) (*store.Deletion
 
 	stats, err := store.DeleteUserEmbeddings(ctx, c.config.PostgresConnStr, doDryRun)
 	if err != nil {
-		slog.Error("Failed to delete embeddings: %v", err)
+		slog.Error("Failed to delete embeddings", "error", err)
 		return nil, err
 	}
 	return stats, nil
@@ -110,7 +110,7 @@ func (c *Client) GetStats(
 ) (*store.GetStatsRow, error) {
 	stats, err := store.GetStats(ctx, c.config.PostgresConnStr)
 	if err != nil {
-		slog.Error("Failed to get user stats: %v", err)
+		slog.Error("Failed to get user stats", "error", err)
 		return nil, err
 	}
 	return stats, nil
