+++ internal/rag/rag_prompt_with_initial_query.txt
